AMD SB900: fix warnings
Add a prototype to a .h file
Remove an unused file (GppHp.c) from the build by deleting it from the
  makefile. I left the file since this is vendorcode. This is the code
  for PCIe hotplug.
Inside GppHp.c, make functions not called from outside static.
  This obviously isn't important since the file isn't used, but for
  the sake of the cleanup I thought I'd go ahead with it...

This was tested with the torpedo build.

This fixes these warnings:

src/vendorcode/amd/cimx/sb900/Dispatcher.c: In function 'LocateImage':
src/vendorcode/amd/cimx/sb900/Dispatcher.c:193:38: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]

src/vendorcode/amd/cimx/sb900/Usb.c:740:1: warning: no previous prototype for 'XhciA12Fix' [-Wmissing-prototypes]

src/vendorcode/amd/cimx/sb900/GppHp.c:65:1: warning: no previous prototype for 'sbGppHotPlugSmiProcess' [-Wmissing-prototypes]
src/vendorcode/amd/cimx/sb900/GppHp.c: In function 'sbGppHotPlugSmiProcess':
src/vendorcode/amd/cimx/sb900/GppHp.c:76:5: warning: implicit declaration of function 'SbStall' [-Wimplicit-function-declaration]
src/vendorcode/amd/cimx/sb900/GppHp.c: At top level:
src/vendorcode/amd/cimx/sb900/GppHp.c:101:1: warning: no previous prototype for 'sbGppHotUnplugSmiProcess' [-Wmissing-prototypes]
src/vendorcode/amd/cimx/sb900/GppHp.c:134:1: warning: no previous prototype for 'sbGppHotplugSmiCallback' [-Wmissing-prototypes]
src/vendorcode/amd/cimx/sb900/GppHp.c: In function 'sbGppHotplugSmiCallback':
src/vendorcode/amd/cimx/sb900/GppHp.c:158:5: warning: implicit declaration of function 'outPort80' [-Wimplicit-function-declaration]

Reviewed-on: http://review.coreboot.org/2349
Tested-by: build bot (Jenkins)

+++ src/vendorcode/amd/cimx/sb900/Dispatcher.c
@@ -190,7 +190,7 @@ LocateImage (
 #ifdef x64
     12346789
 #else
-    Result = VerifyImage (Signature, (VOID*) (__int64)ImagePtr);
+    Result = VerifyImage (Signature, (VOID*) (intptr_t) ImagePtr);
 #endif
     if ( Result != NULL ) {
       return Result;
